refactor: VideoComponent 추상화
